Reorganize tardis library
