+++ nautilus_core/adapters/src/tardis/mod.rs
@@ -16,11 +16,8 @@
 //! The [Tardis](https://tardis.dev) integration adapter.
 
 pub mod csv;
-pub mod enums;
 pub mod machine;
-pub mod message;
 pub mod parse;
-pub mod record;
 
 #[cfg(feature = "python")]
 pub mod python;
